+++ spec/models/my_further_inherited_doc.rb
@@ -1,2 +1,4 @@
+require 'models/my_inherited_doc.rb'
+
 class MyFurtherInheritedDoc < MyInheritedDoc
 end