+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.8.3 (Next)
 
+* [#35](https://github.com/mongoid/mongoid_fulltext/pull/35): Mongoid 7 compatibility - [@tomasc](https://github.com/tomasc).
 * Your contribution here.
 
 ### 0.8.2 (8/5/2018)